backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pandas==2.2.1
python-multipart==0.0.9
gunicorn==21.2.0
python-dotenv==1.0.1
pydantic==2.6.3